refactor: ArticleSection 컴포넌트 폴더 생성 후 정리
