+++ pages/boards/index.tsx
@@ -1,5 +1,5 @@
-import ArticleSection from "@/components/boards/ArticleSection";
-import BestArticleSection from "@/components/boards/BestArticleSection";
+import ArticleSection from "@/components/boards/ArticleSection/ArticleSection";
+import BestArticleSection from "@/components/boards/ArticleSection/BestArticleSection";
 import styles from "@/styles/Boards.module.scss";
 
 function Boards() {
